Extracted proxy to its own file
